refactor: 이미지 업로드 커스텀 훅
