+++ src/pages/Chat/Chat.tsx
@@ -1,6 +1,6 @@
-import { useCallback, useEffect, useRef, useState, ChangeEvent, DragEvent } from "react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { chatResponse } from "../../utils/api.ts";
-import { ChatProps, ContentProps, ImageProps, MessageProps } from "../../types/type.ts";
+import { ChatProps, ContentProps, MessageProps } from "../../types/type.ts";
 import {
   ChatComponent,
   ChatHeader,
@@ -17,12 +17,21 @@ import {
 import Message from "./components/Message/Message.tsx";
 import InputText from "../../components/InputText.tsx";
 import Button from "../../components/Button.tsx";
-
-const MAX_IMAGES = 2;
-const MAX_SIZE = 4;
-const MAX_FILE_SIZE = MAX_SIZE * 1024 * 1024;
+import { useImageUpload } from "../../hook/useImageUpload.tsx";
 
 const Chat = ({ as, className }: ChatProps) => {
+  const {
+    MAX_IMAGES,
+    MAX_SIZE,
+    images,
+    setImages,
+    isDragging,
+    handleDragOver,
+    handleDragEnd,
+    handleDrop,
+    handleFileChange,
+    removeImage,
+  } = useImageUpload();
   const [messages, setMessages] = useState<MessageProps[]>([]);
   const [streamingMessage, setStreamingMessage] = useState<string>("");
   const [input, setInput] = useState<string>("");
@@ -30,8 +39,6 @@ const Chat = ({ as, className }: ChatProps) => {
   const [isInputting, setInputting] = useState<boolean>(false);
   const [isStreaming, setStreaming] = useState<boolean>(false);
   const [inputtingTimer, setInputtingTimer] = useState<NodeJS.Timeout | null>(null);
-  const [images, setImages] = useState<ImageProps[]>([]);
-  const [isDragging, setDragging] = useState<boolean>(false);
 
   const messageEndRef = useRef<HTMLDivElement>(null);
   const fileInputRef = useRef<HTMLInputElement>(null);
@@ -72,17 +79,6 @@ const Chat = ({ as, className }: ChatProps) => {
       reader.onload = () => resolve(reader.result as string);
       reader.onerror = (error) => reject(error);
     });
-  };
-
-  // setImages에 이미지 추가하는 함수
-  const addImages = (files: File[]) => {
-    const imageFiles = files.filter((file) => file.type.startsWith("image/"));
-    const filterFiles = imageFiles.filter((file) => file.size <= MAX_FILE_SIZE);
-    const newImages: ImageProps[] = filterFiles.slice(0, MAX_IMAGES - images.length).map((file) => ({
-      file,
-      preview: URL.createObjectURL(file),
-    }));
-    setImages((prev) => [...prev, ...newImages].slice(0, MAX_IMAGES));
   };
 
   // 메세지를 입력할 때
@@ -176,41 +172,6 @@ const Chat = ({ as, className }: ChatProps) => {
     },
     [isLoading, messages, makeSetMessage, images]
   );
-
-  // 드래그 할 때
-  const handleDragOver = (e: DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    setDragging(true);
-  };
-
-  // 드래그 끝날 때
-  const handleDragEnd = (e: DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    if (!componentRef.current?.contains(e.relatedTarget as Node)) {
-      setDragging(false);
-    }
-  };
-
-  // 드래그 후 놓을 때
-  const handleDrop = (e: DragEvent<HTMLDivElement>) => {
-    e.preventDefault();
-    setDragging(false);
-    const files = Array.from(e.dataTransfer.files);
-    addImages(files);
-  };
-
-  // 이미지 선택
-  const handleFileChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const files = e.target.files;
-    if (files) {
-      addImages([...files]);
-    }
-  };
-
-  // 이미지 삭제
-  const removeImage = (index: number) => {
-    setImages((prev: ImageProps[]) => prev.filter((_, idx) => idx !== index));
-  };
 
   return (
     <ChatComponent
